+++ next.config.ts
@@ -4,4 +4,15 @@ const nextConfig: NextConfig = {
   /* config options here */
 };
 
+module.exports = {
+    async rewrites() {
+      return [
+        {
+          source: '/@:username/:slug*',
+          destination: '/:username/:slug*'
+        }
+      ]
+    }
+  }
+
 export default nextConfig;
