+++ src/lib/api/article/read/type.ts
@@ -1,0 +1,14 @@
+interface ReadArticleResult {
+    id: string
+    authorId: string
+    createdAt: string
+    updatedAt: string
+    folderId: string,
+    summary: string,
+    thumbnailUrl: string | null,
+    isPublic : string,
+    title: string,
+    content: string
+}
+
+
