+++ src/lib/api/auth/refresh.ts
@@ -1,7 +1,7 @@
 import { store } from "@/app/store/store"
 import { setAccessTokenAction } from "@/features/auth/store/authSlice"
 import { setUser } from "@/features/user/store/userSlice"
-import { defaultFetch } from "@/lib/customFetch"
+import { defaultFetch } from "@/lib/fetch/defaultFetch"
 
 export const refreshAccessToken = async () : Promise<AuthResult> => {
     try {
