+++ src/lib/fetch/ssrFetch.ts
@@ -1,0 +1,19 @@
+import { cookies } from "next/headers";
+import { FetchOptions, defaultFetch } from "./defaultFetch";
+
+export const ssrFetch = async <T = any>(
+    uri: string, options: FetchOptions
+) : Promise<T | undefined>  => {
+    const cookieStore = await cookies()
+    const accessToken = cookieStore.get('access_token')?.value
+
+    try {
+        const authHeaders = new Headers(options.headers)
+        authHeaders.set('Authorization', `Bearer ${accessToken}`)
+
+        return await defaultFetch<T>(uri, {...options, headers: authHeaders})
+    }
+    catch (err) {
+        throw err
+    }
+}
